# Conflicts:
#   src/app/core/services/auth.service.ts
#   src/app/profile/profile-overview/profile-overview.component.ts

+++ src/app/shared/tree/tree.component.ts
@@ -1,10 +1,12 @@
-import {Component, ElementRef, EventEmitter, Input, OnChanges, Output, SimpleChanges, ViewChild} from '@angular/core';
+import {Component, ElementRef, EventEmitter, Input, OnChanges, OnInit, Output, SimpleChanges, ViewChild} from '@angular/core';
 import {FlatTreeControl} from '@angular/cdk/tree';
 import {MatTreeFlatDataSource, MatTreeFlattener} from '@angular/material';
 import {SelectionModel} from '@angular/cdk/collections';
-import {InsertionType, TreeDatabaseService} from './providers/tree-database.service';
+import {InsertionType, TreeDatabaseService} from './services/tree-database.service';
 import {ItemFlatNode, ItemNode} from './models/item-node.model';
 import {Observable, of} from 'rxjs';
+import {DeleteProtege} from '../../root-store/mentors/mentors.actions';
+import {DialogService} from '../dialog/services/dialog.service';
 
 @Component({
   selector: 'lt-tree',
@@ -13,7 +15,7 @@ import {Observable, of} from 'rxjs';
   providers: [TreeDatabaseService]
 })
 
-export class TreeComponent implements OnChanges {
+export class TreeComponent implements OnChanges, OnInit {
   @Input() private data: ItemNode[];
   @Input() public canEdit = true;
   @Input() public type: new (...arg: any[]) => ItemNode;
@@ -49,7 +51,7 @@ export class TreeComponent implements OnChanges {
 
   saveIsDisabled = true;
 
-  constructor(private database: TreeDatabaseService) {
+  constructor(private database: TreeDatabaseService, private dialogService: DialogService) {
     this.treeFlattener = new MatTreeFlattener(this.transformer, this.getLevel, this.isExpandable, this.getChildren);
     this.treeControl = new FlatTreeControl<ItemFlatNode>(this.getLevel, this.isExpandable);
     this.dataSource = new MatTreeFlatDataSource(this.treeControl, this.treeFlattener);
@@ -58,10 +60,14 @@ export class TreeComponent implements OnChanges {
 
     database.dataChange
       .subscribe(data => {
-      this.dataSource.data = [];
-      this.dataSource.data = data;
-      this.dataChanged.emit(data);
-    });
+        this.dataSource.data = [];
+        this.dataSource.data = data;
+        this.dataChanged.emit(data);
+      });
+  }
+
+  ngOnInit(): void {
+    this.treeControl.expandAll();
   }
 
   getLevel = (node: ItemFlatNode) => node.level;
@@ -181,9 +187,24 @@ export class TreeComponent implements OnChanges {
   removeItem(node: ItemFlatNode) {
     const data = this.flatNodeMap.get(node);
     const parent = this.database.getParentFromNodes(data);
-    this.database.deleteItem(data);
-    if (node.showAsInput !== 'add') {
-      this.deleteItem.emit(data);
+    console.log(node);
+    if (node.showAsInput !== 'add' && node.showAsInput !== 'edit') {
+      const htmlContent = `<p>Вы уверены, что хотите удалить пункт <b>${node.text}</b> ?</p>`;
+      this.dialogService.openConfirmDialog({ htmlContent }, (confirm) => {
+        if (confirm) {
+          this.database.deleteItem(data);
+          this.deleteItem.emit(data);
+        }
+      });
+    }
+
+    if (node.showAsInput === 'add') {
+      this.database.deleteItem(data);
+    }
+
+    if (node.showAsInput === 'edit') {
+      data.showAsInput = false;
+      this.database.updateItem(data, data.text);
     }
   }
